Improve race condition of setting the controlling terminal
Switching VTs can take a long time and sleep, which gives the child
process a chance to start and possibily exit before we ever set the
controlling terminal. This caused an error message to be printed for
short running processes, like echo.

The race condition is improved here, but theoretically, there could
still be a problem if the child process tries to do something that
requires the controlling terminal to be set right away.